mod uv_g

+++ tamfunc/tamfunc.py
@@ -517,7 +517,7 @@ def w_area(da,ndims=3,lataxis=1):
 #
 def uv_g(ssh_da,):
     ssh = ssh_da.values.astype(np.float32)
-    lat = np.deg2rad(ssh_da.lat.values)
+    lat = np.deg2rad(ssh_da.lat.values).astype(np.float32)
     lon = ssh_da.lon.values.astype(np.float32)
     dlon = np.deg2rad(lon[1]-lon[0])
     dlat = lat[1]-lat[0]
